rename template.yaml resources. directory name app_stack -> app update readme.md file tree
