+++ tests/unit/test_handler.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from app_stack import hello_world
+from app import get_message
 
 
 @pytest.fixture()
@@ -63,7 +63,7 @@ def apigw_event():
 
 
 def test_lambda_handler(apigw_event, mocker):
-    ret = hello_world.lambda_handler(apigw_event, "")
+    ret = get_message.lambda_handler(apigw_event, "")
     data = json.loads(ret["body"])
 
     assert ret["statusCode"] == 200
